backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
httpx
pydantic[email]
email-validator
redis
limits
slowapi
apscheduler>=3.10.0

# Data processing
pandas
numpy
pytz

# Exchange integration
ccxt>=4.0.0

# Security
cryptography
PyJWT>=2.8.0

# AI Coach
openai>=1.3.0
tiktoken>=0.5.0
anthropic>=0.40.0

# Supabase
supabase>=2.0.0

# Form handling
python-multipart
